fix: resolve linting and build errors
This commit addresses several linting and build errors that were causing the CI to fail.

- Resolves `react-hooks/set-state-in-effect` and `react/jsx-no-undef` in `app/client/connect/page.tsx` by refactoring the component to be client-side only and initializing state from cookies.
- Fixes `react-hooks/immutability` in `hooks/useAutoConnect.ts` by refactoring the hook to use a stable `useCallback` function.
- Corrects `no-undef` in `jest.config.cjs` by adding an `eslint-env` comment.
- Removes unused variables in `services/tabataTimer.ts` and `tests/playwright/mobile-assessment.spec.ts`.
- Fixes a TypeScript error in `hooks/useBluetoothHRM.ts` by correctly checking the device's connection status.
- Resolves a prerendering error in `app/client/connect/page.tsx` by dynamically importing the component.
