+++ hooks/useAutoConnect.ts
@@ -9,51 +9,47 @@ type ConnectFn = () => Promise<boolean>
 const useAutoConnect = (connectFn: ConnectFn, start: boolean) => {
   const [isConnecting, setIsConnecting] = useState(false)
   const [attempts, setAttempts] = useState(0)
-  const [delay, setDelay] = useState(INITIAL_DELAY)
   const timeoutRef = useRef<NodeJS.Timeout | null>(null)
+  const delayRef = useRef(INITIAL_DELAY)
+  const savedConnectFn = useRef(connectFn)
 
-  const tryConnect = useCallback(async () => {
-    setIsConnecting(true)
-    setAttempts((prev) => prev + 1)
+  useEffect(() => {
+    savedConnectFn.current = connectFn
+  }, [connectFn])
 
-    const success = await connectFn()
-
-    if (success) {
-      setIsConnecting(false)
-      // Reset attempts and delay on success
-      setAttempts(0)
-      setDelay(INITIAL_DELAY)
-    } else {
-      // Exponential backoff
-      const newDelay = Math.min(delay * 2, MAX_DELAY)
-      setDelay(newDelay)
-
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current)
+  const tryConnect = useCallback(() => {
+    const connect = async () => {
+      setIsConnecting(true)
+      setAttempts((prev) => prev + 1)
+      const success = await savedConnectFn.current()
+      if (success) {
+        setIsConnecting(false)
+        setAttempts(0)
+        delayRef.current = INITIAL_DELAY
+      } else {
+        const newDelay = Math.min(delayRef.current * 2, MAX_DELAY)
+        delayRef.current = newDelay
+        if (timeoutRef.current) {
+          clearTimeout(timeoutRef.current)
+        }
+        timeoutRef.current = setTimeout(connect, newDelay)
       }
-
-      timeoutRef.current = setTimeout(tryConnect, newDelay)
     }
-  }, [connectFn, delay])
+    connect()
+  }, [])
 
   useEffect(() => {
     if (start) {
-      // Start the connection process
       tryConnect()
-    } else {
-      // Stop and reset everything if start becomes false
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current)
-      }
-      setIsConnecting(false)
-      setAttempts(0)
-      setDelay(INITIAL_DELAY)
     }
 
     return () => {
       if (timeoutRef.current) {
         clearTimeout(timeoutRef.current)
       }
+      setIsConnecting(false)
+      setAttempts(0)
+      delayRef.current = INITIAL_DELAY
     }
   }, [start, tryConnect])
